src/fastapi/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html 
torch==2.0.0+cpu
-f https://download.pytorch.org/whl/torch_stable.html 
torchvision==0.15.1+cpu

tiktoken
fastapi[all]==0.98.0
Pillow==9.5.0
gunicorn==20.1.0